경고나는 코드 수정.
형변환 관련 코드가 형병환 경고를 발생하여,
관련된 코드를 수정함.

+++ TimeAttack/20140512/BaseType.cpp
@@ -29,11 +29,11 @@ Size::Size(const LONG& _cx/*=0*/, const LONG& _cy/*=0*/)
 }
 Size operator * (const float& v, const Size& obj)
 {
-	return Size(obj.cx*v, obj.cy*v);
+	return Size(LONG(obj.cx*v), LONG(obj.cy*v));
 }
 Size operator * (const Size& obj, const float& v)
 {
-	return Size(obj.cx*v, obj.cy*v);
+	return Size(LONG(obj.cx*v), LONG(obj.cy*v));
 }
 Size operator * (const LONG& v, const Size& obj)
 {
